moved all api helpers to agent
